requirements.txt
# Core dependencies
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Data processing
pandas>=2.0.0

# Text processing and similarity matching
rapidfuzz>=3.0.0
regex>=2023.0.0

# Progress tracking
tqdm>=4.65.0

# Type hints (for development)
typing-extensions>=4.5.0 